Fixed permissions bug
Login would fail if an user had no permissions due to a faulty SQL join.
The INNER JOIN operation returns an empty result if one table has no
entries for a specified targed. Therefore it was necessary to use
LEFT JOIN so the results of the left table are still returned even if
the right table has no results